content/base/test/test_bug456262.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=456262
-->
<head>
  <title>Test for Bug 456262</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=456262">Mozilla Bug 456262</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 456262 **/

function runTest() {
  document.expando = document;
  document.documentElement.expando = document;
  document.documentElement.setAttribute("foo", "bar");
  document.documentElement.getAttributeNode("foo").expando = document;
  SpecialPowers.gc();
  ok(document.expando, "Should have preserved the expando!");
  ok(document.documentElement.expando, "Should have preserved the expando!");
  ok(document.documentElement.getAttributeNode('foo').expando,
     "Should have preserved the expando!");
}

runTest();

</script>
</pre>
</body>
</html>
